+++ test/src/models/form_control_test.dart
@@ -87,5 +87,10 @@ void main() {
 
       expect(formControl.value, defaultValue);
     });
+
+    test('Assert error if debounce time < 0', () {
+      final formControl = () => FormControl(asyncValidatorsDebounceTime: -1);
+      expect(formControl, throwsAssertionError);
+    });
   });
 }